--- app/skill/models/__init__.py
@@ -1,1 +1,0 @@
-from .skill import *
